requirements.txt
# Polymarket Sports Betting Bot Dependencies
web3>=6.0.0
eth-account>=0.5.0
requests>=2.28.0
aiohttp>=3.8.0
websockets>=12.0
pandas>=2.0.0
xgboost>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
